Charts help files refactor
- renamed help files to match node full (type) names
