--- doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.BarChart.md
@@ -1,7 +1,0 @@
-## In Depth
-Bar Chart text.
-___
-## Example File
-
-![Bar Chart](./CoreNodeModelsWpf.Charts.BarChart_img.jpg)
-
