+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.BarChartNodeModel.dyn
