+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.BarChartNodeModel.md
@@ -1,0 +1,7 @@
+## In Depth
+Bar Chart text.
+___
+## Example File
+
+![Bar Chart](./CoreNodeModelsWpf.Charts.BarChartNodeModel_img.jpg)
+
