+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.BarChartNodeModel_img.jpg
