+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.BasicLineChartNodeModel.dyn
