+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.BasicLineChartNodeModel.md
@@ -1,0 +1,7 @@
+## In Depth
+Index-Value Line Plot text.
+___
+## Example File
+
+![Index-Value Line Plot](./CoreNodeModelsWpf.Charts.BasicLineChartNodeModel_img.jpg)
+
