+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.HeatSeriesNodeModel.dyn
