+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.HeatSeriesNodeModel.md
@@ -1,0 +1,7 @@
+## In Depth
+Heat Series Plot text.
+___
+## Example File
+
+![Heat Series Plot](./CoreNodeModelsWpf.Charts.HeatSeriesNodeModel_img.jpg)
+
