+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.HeatSeriesNodeModel_img.jpg
