--- doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.HeatSeriesPlot.md
@@ -1,7 +1,0 @@
-## In Depth
-Heat Series Plot text.
-___
-## Example File
-
-![Heat Series Plot](./CoreNodeModelsWpf.Charts.HeatSeriesPlot_img.jpg)
-
