--- doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.Index-ValueLinePlot.md
@@ -1,7 +1,0 @@
-## In Depth
-Index-Value Line Plot text.
-___
-## Example File
-
-![Index-Value Line Plot](./CoreNodeModelsWpf.Charts.Index-ValueLinePlot_img.jpg)
-
