--- doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.PieChart.md
@@ -1,7 +1,0 @@
-## In Depth
-Pie Chart text.
-___
-## Example File
-
-![Pie Chart](./CoreNodeModelsWpf.Charts.PieChart_img.jpg)
-
