+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.PieChartNodeModel.md
@@ -1,0 +1,7 @@
+## In Depth
+Pie Chart text.
+___
+## Example File
+
+![Pie Chart](./CoreNodeModelsWpf.Charts.PieChartNodeModel_img.jpg)
+
