+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.PieChartNodeModel_img.jpg
