--- doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.ScatterPlot.md
@@ -1,7 +1,0 @@
-## In Depth
-Scatter Plot text.
-___
-## Example File
-
-![Scatter Plot](./CoreNodeModelsWpf.Charts.ScatterPlot_img.jpg)
-
