+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.ScatterPlotNodeModel.dyn
