+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.ScatterPlotNodeModel.md
@@ -1,0 +1,7 @@
+## In Depth
+Scatter Plot text.
+___
+## Example File
+
+![Scatter Plot](./CoreNodeModelsWpf.Charts.ScatterPlotNodeModel_img.jpg)
+
