+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.ScatterPlotNodeModel_img.jpg
