+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.XYLineChartNodeModel.dyn
