+++ doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.XYLineChartNodeModel.md
@@ -1,0 +1,7 @@
+## In Depth
+XY Line Plot text.
+___
+## Example File
+
+![XY Line Plot](./CoreNodeModelsWpf.Charts.XYLineChartNodeModel_img.jpg)
+
