--- doc/distrib/NodeHelpFiles/CoreNodeModelsWpf.Charts.XYLinePlot.md
@@ -1,7 +1,0 @@
-## In Depth
-XY Line Plot text.
-___
-## Example File
-
-![XY Line Plot](./CoreNodeModelsWpf.Charts.XYLinePlot_img.jpg)
-
