+++ src/Libraries/CoreNodeModelsWpf/Charts/Controls/BasicLineChartControl.xaml
@@ -1,9 +1,9 @@
-<UserControl x:Class="CoreNodeModelsWpf.Charts.Controls.BasicLineChartControl"
+<UserControl x:Class="CoreNodeModelsWpf.Charts.BasicLineChartControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CoreNodeModelsWpf.Charts.Controls"
+             xmlns:local="clr-namespace:CoreNodeModelsWpf.Charts"
              xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
              xmlns:ui="clr-namespace:Dynamo.UI;assembly=DynamoCoreWpf"
              mc:Ignorable="d"
